requirements.txt
Django==2.2
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
